refactor: 薄いラッパークラスとデッドコードを削除
## 背景
PR #1319 で crane_session_coordinator の薄いラッパークラス3つを統合したのと
同様に、コードベース全体で「1箇所からしか使われない薄いラッパー」と
「どこからも使われないデッドコード」を整理し、コードの見通しを改善する。

## 変更内容
- `Hysteresis`（crane_game_analyzer）: コードベース全体でどこからも参照されない
  デッドコードを削除
- `BallPredictionTracker`（crane_msg_wrappers）: 外部から参照ゼロのデッドコード
  （hpp+cpp）を削除
- `visualization_helpers.hpp`（crane_local_planner）: `drawRobotRadiusWithSpeed()`
  を rvo2_planner.cpp の無名namespace内にファイルローカル関数として統合し削除
- `KickPowerCalculator` struct（crane_local_planner）: `LocalPlannerComponent` に
  直接統合。`kicker_model_` メンバと `getKickPower()` メソッドとして吸収
- `BroadcastCommandSender`（crane_sender）: `IbisSenderNode` のブロードキャスト
  ソケット操作ロジックを直接展開。UDPSender 系メンバと同じ扱いに統一し削除

削除ファイル数: 計7ファイル（hpp 5 + cpp 2）
