+++ crane_local_planner/include/crane_local_planner/local_planner.hpp
@@ -25,61 +25,6 @@
 namespace crane
 {
 
-struct KickPowerCalculator
-{
-private:
-  // KickerModel統合（すべてのキック計算用）
-  std::shared_ptr<KickerModel> kicker_model;
-
-public:
-  auto initialize(rclcpp::Node & node) -> void
-  {
-    // KickerModel初期化（YAML設定ファイルから読み込み）
-    node.declare_parameter<std::string>("kicker_physics_config", "");
-    std::string kicker_config_path = node.get_parameter("kicker_physics_config").as_string();
-
-    try {
-      if (!kicker_config_path.empty()) {
-        // YAML設定ファイルからKickerModelを作成
-        kicker_model = createKickerModelFromYAML(kicker_config_path);
-        RCLCPP_INFO(
-          node.get_logger(), "KickerModel initialized from YAML: %s", kicker_config_path.c_str());
-      } else {
-        // デフォルト設定でKickerModelを作成
-        kicker_model = createDefaultKickerModel();
-        RCLCPP_INFO(node.get_logger(), "KickerModel initialized with default values");
-      }
-    } catch (const std::exception & e) {
-      RCLCPP_ERROR(node.get_logger(), "KickerModel initialization failed: %s", e.what());
-      throw std::runtime_error("Failed to initialize KickerModel: " + std::string(e.what()));
-    }
-  }
-
-  auto getKickPower(const crane_msgs::msg::RobotCommand & command) const -> double
-  {
-    if (not command.local_planner_config.kick_power_override) {
-      return command.kick_power;
-    }
-
-    // KickerModelを使用（必須）
-    if (!kicker_model) {
-      throw std::runtime_error("KickerModel is not initialized");
-    }
-
-    try {
-      if (command.chip_enable) {
-        return kicker_model->calculateChipKickPower(
-          command.local_planner_config.target_chip_distance);
-      } else {
-        return kicker_model->calculateStraightKickPower(
-          command.local_planner_config.target_kick_ball_speed);
-      }
-    } catch (const std::exception & e) {
-      throw std::runtime_error("KickerModel calculation failed: " + std::string(e.what()));
-    }
-  }
-};
-
 class LocalPlannerComponent : public rclcpp::Node
 {
 public:
@@ -94,7 +39,22 @@ public:
     declare_parameter("planner", "rvo2");
     auto planner_str = get_parameter("planner").as_string();
 
-    kick_power_calculator.initialize(*this);
+    // KickerModel初期化（YAML設定ファイルから読み込み）
+    declare_parameter<std::string>("kicker_physics_config", "");
+    std::string kicker_config_path = get_parameter("kicker_physics_config").as_string();
+    try {
+      if (!kicker_config_path.empty()) {
+        kicker_model_ = createKickerModelFromYAML(kicker_config_path);
+        RCLCPP_INFO(
+          get_logger(), "KickerModel initialized from YAML: %s", kicker_config_path.c_str());
+      } else {
+        kicker_model_ = createDefaultKickerModel();
+        RCLCPP_INFO(get_logger(), "KickerModel initialized with default values");
+      }
+    } catch (const std::exception & e) {
+      RCLCPP_ERROR(get_logger(), "KickerModel initialization failed: %s", e.what());
+      throw std::runtime_error("Failed to initialize KickerModel: " + std::string(e.what()));
+    }
 
     crane::CraneVisualizerBuffer::activate(*this);
 
@@ -137,7 +97,7 @@ private:
 
   double theta_offset = 0.;
 
-  KickPowerCalculator kick_power_calculator;
+  std::shared_ptr<KickerModel> kicker_model_;
 
   DiagnosticHelper diagnostic_helper_;
 
@@ -148,6 +108,27 @@ private:
 
   auto aggregateStates(const std::vector<crane_msgs::msg::NamedString> & planning_factors) const
     -> std::string;
+
+  auto getKickPower(const crane_msgs::msg::RobotCommand & command) const -> double
+  {
+    if (not command.local_planner_config.kick_power_override) {
+      return command.kick_power;
+    }
+    if (!kicker_model_) {
+      throw std::runtime_error("KickerModel is not initialized");
+    }
+    try {
+      if (command.chip_enable) {
+        return kicker_model_->calculateChipKickPower(
+          command.local_planner_config.target_chip_distance);
+      } else {
+        return kicker_model_->calculateStraightKickPower(
+          command.local_planner_config.target_kick_ball_speed);
+      }
+    } catch (const std::exception & e) {
+      throw std::runtime_error("KickerModel calculation failed: " + std::string(e.what()));
+    }
+  }
 };
 
 }  // namespace crane
